requirements.txt
telethon>=1.24.0
pandas>=1.3.0
plotly>=5.3.0
python-dateutil>=2.8.2
kaleido>=0.2.1
google-generativeai>=0.3.0
openai>=1.10.0
python-dotenv>=0.21.0
networkx>=2.8.0
scikit-learn>=1.0.0
wordcloud>=1.9.0
gensim>=4.0.0